Export mutations from a single file
This follows up on the good idea of @peacheym by collapsing the files together
and exporting everything from there.
