--- frontend/src/graphql/mutations/validateTwoFactor.js
@@ -1,13 +1,0 @@
-import gql from 'graphql-tag'
-
-const VALIDATE_TWO_FACTOR = gql`
-  mutation ValidateTwoFactor($userName: EmailAddress!, $otpCode: String!) {
-    authenticateTwoFactor(userName: $userName, otpCode: $otpCode) {
-      user {
-        userName
-      }
-    }
-  }
-`
-
-export default VALIDATE_TWO_FACTOR
